20.1.The Magic '_controller' Attribute {Overriding the Controller from a Listener?; Overriding the Controller on kernel.request?;Callback Controller with _controller}
Now that we've been stuffed full of knowledge about the request-response process, let's see what kind of trouble we can get into. Uh, I mean, let's do some cool and productive things with all this new information!

Overriding the Controller from a Listener?
Close all the files except for index.php and HttpKernel. Here's our first challenge: could we - from some event listener - change the controller for the page?
Hint: in Symfony, the answer to "can I do X" is always yes. In this case, it's not only possible, there are multiple ways to do it.
For example, when HttpKernel dispatches the kernel.controller event, it passes each listener a ControllerEvent object. And that class has a setController() method. Easy peasy! We can override the controller by adding a listener to that event. Heck, you can do the same thing down here with the kernel.controller_arguments event.

Overriding the Controller on kernel.request?
So... that was too easy. I'll close up my directory tree... and then open our UserAgentSubscriber. Here's my harder challenge: how can we override the controller from here: from a listener to the kernel.request event. In this case, there is no setController() method.

Callback Controller with _controller
The trick is to remember how the controller resolver works: it starts by fetching the _controller value from the $request->attributes. So if, for some reason, we wanted to completely replace the controller, we can do it right here: $request->attributes->set('_controller', .... For fun, let's set this to an anonymous function... cause yea! That's allowed! Inside, return a new Response() with:

I just took over the controller

Will it work? Refresh any page. Yep! We see our message here, on the homepage and everywhere.