20.2.The Magic '_controller' Attribute {Callback Controller with _controller; RouterListener Skips when _controller is Set}
Our normal controller tricks work just fine: add a $slug argument... but give it a default value and then dd($slug). On the article show page... this works thanks to the {slug} wildcard. On the homepage, it's null because that wildcard doesn't exist.

RouterListener Skips when _controller is Set
Open up RouterListener.php one more time and find its onKernelRequest() method. This method is normally responsible for executing the router and setting the _controller key onto the request attributes. But back down at getSubscribedEvents(), ah! The kernel.request listener has a priority of 32. We didn't give our subscriber a priority, so it has a priority of 0. This means that RouterListener is called before our subscriber.
Ok, so then here is what's happening: RouterListener is called first and it is executing the router and setting an _controller key on the request attributes. Then our listener is called and we override that value.
So... if we reversed the order - if we made our listener be called first - our little _controller hack would not work, right? Because RouterListener would override our value.
Actually... no! At the top of onKernelRequest(), one of the first things it does is check to see if something has already set the _controller attribute. If it has, it does nothing: someone else has decided to be responsible for figuring out which controller to call. In reality, no matter how early the _controller attribute is set, it will always win over RouterListener.
Why is that important? Because this explains how ErrorListener was able to execute ErrorController. Open up ErrorListener.php. Remember: to execute ErrorController this duplicated the request. But it didn't create an exact copy: it overrode the attributes in order to set _controller to error_controller. Then it sent that new Request back through through the entire $kernel->handle() process! This means that before any listeners were executed during that second trip through HttpKernel::handle(), the _controller attribute was already set.