20.3.The Magic '_controller' Attribute {Peaking at our First Sub-Request}
Peaking at our First Sub-Request

So in reality, on an error page, RouterListener is called two times: once for the main request... when it does its job normally... and again for the "sub request". That second time, because the _controller attribute is already set, RouterListener does nothing.
In fact, let's see this. Before the if, dump($request->attributes->has('_controller')). Then, in your browser, go back to a 404 and try it.
Ah, boo! This hit the die statement in our fake controller. I didn't mean to do that. In UserAgentSubscriber, comment-out our controller hack so we can see the whole process.
Ok, try it again. Hello 404 page! Hover over the target icon on the web debug toolbar. Yes! 2 dumps from RouterListener: false the first time it's called and, the second time it's called - which is due to the code in ErrorListener - it dumps true because that Request does already have the _controller attribute.
This second request is called a sub-request... but more on that topic later. Remove the dump() call.
Let's see what other ways we can hack into Symfony, like by adding new things that can be used as controller arguments. That's next.