21.1.Custom Global Controller Arguments
Now that we understand a lot more about its flow, we're on a mission to find weird, crazy things that we can do in Symfony. For the next one, pretend that, for some reason, we need to know whether or not a visitor is using a Mac or not. In fact, we need this info so often, that we want the ability to add an $isMac argument to any controller, like this.

Let's dump()... and then try it. No surprise, it explodes!

  Controller show() requires that you provide a value for the  argument.

I'll go back to a real article page, though that won't make any difference.