21.2.Custom Global Controller Arguments {Custom Arguments Via Request Attributes}
Custom Arguments Via Request Attributes
So: how can we make this work? There are actually two answers, and we're going to try both. The first is a, kind of, lower-level way of doing it. We know that if we have a {slug} route wildcard, we are allowed to have a $slug argument. So, in theory, if we had an {isMac} wildcard, we could have an $isMac argument, though that's not what we want.
But it's not really that we're allowed to have a $slug argument because there's a {slug} wildcard. Nope, we're allowed to have a $slug argument because there is a slug key in the $request->attributes. The router puts slug into attributes because of the wildcard, but when it comes to figuring out what arguments to pass to a controller, it's all about the $request->attributes.
Inside of our listener, let's say $isMac = stripos($userAgent, 'Mac') !== false. Now, to make isMac available as an argument to any controller, add $request->attributes->set('isMac', $isMac).
And... that's it! Try the page now. It works! And for me, it's set to true.