21.3.Custom Global Controller Arguments {Custom ArgumentValueResolver}
Custom ArgumentValueResolver
The second way to add a custom controller argument is a bit more direct: create a custom ArgumentValueResolver. When we were deep-diving into how Symfony determines what arguments to pass to a controller, we found out that there are various classes that determine this called "argument value resolvers". And we can create our own.
Inside of the src/ directory - it doesn't matter, let's put it in Service/ - create a new class called: IsMacArgumentValueResolver. The only rule is that this class must implement ArgumentValueResolveInterface. I'll go to the Code -> Generate menu - or Command + N on a Mac - and select "Implement Methods" to generate the two methods that we need.
Without doing anything else, this class is already being used by Symfony as an argument value resolver. When we talked about that system, I hinted that the way you get an argument value resolver into the system is by creating a service and tagging it with controller.argument_value_resolver. Find your terminal and, once again, run:

php bin/console debug:container --tag=controller.argument_value_resolver
And now... if you look at the service ids, one of them is for our App\Service\IsMacArgumentValueResolver. It's wrapped in another class because Symfony is decorating the services with TraceableValueResolver, but this is our service being used. Our new service already has the tag thanks to Symfony's auto-configuration feature.