21.4.Custom Global Controller Arguments {Custom ArgumentValueResolver}
Filling in the ArgumentValueResolver Logic
Let's go fill in the logic. Here's the plan: very simply, if the argument's name exactly matches , we'll fill in our value. So for supports(), return $argument->getName() === 'isMac'.
For resolve(), go grab the  code from the subscriber, paste it, and then also copy the stripos() logic. Delete the last two lines from the subscriber so that it stops setting this global argument.
Finish up the resolver by saying return stripos($userAgent, 'Mac') !== false.
Let's try it! Find your browser, refresh and.. boo!

  Can use "yield from" only with arrays and Traversables

That's a funny way of saying that I forgot to yield instead of return from this method: resolve() returns a traversable. Try it now and... it works! We still see true for the dump.
Next, let's uncover one last mystery about controller arguments. Back in ArticleController::show(), we originally had an  argument that was type-hinted with an Article entity class. How did that work? Who was making that automatic query for us?