24.1.Sub Requests & Request Data
Remember that cool trick that we did a few minutes ago with the argument value resolver that allowed us to have an $isMac argument to any controller in our system? Does that also work in a controller that's called by sub request? Of course! Because there's nothing special about this controller: it was called thanks to a complete cycle through HttpKernel::handle(). All the same listeners and all the same argument value resolvers are called.
So... cool! Let's use that! Add an $isMac argument... then pass it into the template.
Inside trendingQuotes.html.twig, near the bottom, add {% if isMac %}{% endif %} and inside, put an <hr>, a <small> tag, and then say:

  BTW, you're using a Mac!

Easy enough! Find your browser... navigate back to the homepage... and refresh. On the right, there it is! We're using a Mac.